- Resolved various Visual C++ compiler warnings.

+++ UniPCemu/cpu/opcodes_386.c
@@ -4412,7 +4412,7 @@ uint_32 op_grp2_32(byte cnt, byte varshift)
 		{
 			FLAGW_CF(s); //Save LSB!
 			s = ((s >> 1)&0x7FFFFFFFULL) | ((uint_64)FLAG_CF << 31);
-			overflow = ((s >> 31) ^ ((s >> 30) & 1));
+			overflow = (byte)((s >> 31) ^ ((s >> 30) & 1U));
 		}
 		if (maskcnt && (numcnt==0)) FLAGW_CF(s>>31); //Always sets CF, according to various sources?
 		if (maskcnt) FLAGW_OF(overflow); //Overflow?
@@ -4460,7 +4460,7 @@ uint_32 op_grp2_32(byte cnt, byte varshift)
 			FLAGW_CF(s>>31);
 			//if (s & 0x8) FLAGW_AF(1); //Auxiliary carry?
 			s = (s << 1) & 0xFFFFFFFFU;
-			overflow = (FLAG_CF^(s>>31));
+			overflow = (byte)(FLAG_CF^(s>>31));
 		}
 		if (maskcnt && (numcnt==0)) FLAGW_CF(s>>31); //Always sets CF, according to various sources?
 		if (maskcnt) FLAGW_OF(overflow);
@@ -4475,7 +4475,7 @@ uint_32 op_grp2_32(byte cnt, byte varshift)
 		overflow = numcnt?0:FLAG_OF;
 		for (shift = 1; shift <= numcnt; shift++)
 		{
-			overflow = (s>>31);
+			overflow = (byte)(s>>31);
 			FLAGW_CF(s);
 			//backup = s; //Save backup!
 			s = s >> 1;
